requirements.txt
fastapi
motor
dnspython
python-dotenv
google-genai
pydantic
crawl4ai
uvicorn
httpx